+++ src/algorithm/pervade.rs
@@ -1463,7 +1463,7 @@ pub mod div {
     }
 }
 
-pub mod modulus {
+pub mod modulo {
     use super::*;
     pub fn num_num(a: f64, b: f64) -> f64 {
         b.rem_euclid(a).abs()