+++ src/algorithm/zip.rs
@@ -354,7 +354,7 @@ fn prim_dy_fast_fn(
         Mul => spanned_dy_fn(span, Value::mul),
         Div => spanned_dy_fn(span, Value::div),
         Pow => spanned_dy_fn(span, Value::pow),
-        Modulus => spanned_dy_fn(span, Value::modulus),
+        Modulo => spanned_dy_fn(span, Value::modulo),
         Log => spanned_dy_fn(span, Value::log),
         Eq => spanned_dy_fn(span, Value::is_eq),
         Ne => spanned_dy_fn(span, Value::is_ne),